Re-arrange authorization for updating/inserting content and duplicate the auth data in the request
Previously the authorization check was done at the last moment before updating the database. Now it's done before the field data is parsed out, which means the authz module won't have access to that data.
Ultimately we'll do authorization based on individual fields and the content type so this needs to be fixed.