Move httpd.js to the main directory
Since the httpd.js seems to be fairly generic, I think we can move it into the repository root.

Still to do, add functionality to copy files from setup/* to the site root as necessary.

The only bad part is that httpd.js is only designed to support a single site.